+++ tests/Feature/DependencyInjection/DependencyInjectionNavigationTest.php
@@ -17,6 +17,7 @@ use Symfony\Lsp\Feature\DependencyInjection\DependencyInjectionReferencesHandler
 use Symfony\Lsp\Feature\DependencyInjection\DependencyInjectionSourceFacts;
 use Symfony\Lsp\Feature\DependencyInjection\DependencyInjectionSourceIndexRegistry;
 use Symfony\Lsp\Feature\DependencyInjection\DependencyInjectionSymbolResolver;
+use Symfony\Lsp\Feature\DependencyInjection\ParameterExpressionScanner;
 use Symfony\Lsp\Feature\DependencyInjection\ParameterIndexRegistry;
 use Symfony\Lsp\Feature\DependencyInjection\PhpAutowireReferenceExtractor;
 use Symfony\Lsp\Feature\DependencyInjection\PhpClassDeclaration;
@@ -91,14 +92,14 @@ final class DependencyInjectionNavigationTest extends TestCase
         $yamlExtractor = new YamlDependencyInjectionExtractor(
             new YamlDocumentParser(new NativeTreeSitterParser(new TreeSitterResultDecoder())),
             new YamlDependencyInjectionDeclarationExtractor($converter),
-            new YamlDependencyInjectionReferenceExtractor($converter),
+            new YamlDependencyInjectionReferenceExtractor($converter, new ParameterExpressionScanner()),
         );
         $phpParser = new TolerantPhpParser(new Parser());
-        $autowireExtractor = new PhpAutowireReferenceExtractor($converter, $phpParser);
+        $autowireExtractor = new PhpAutowireReferenceExtractor($converter, $phpParser, new ParameterExpressionScanner());
         $classExtractor = new PhpClassDeclarationExtractor($converter, $phpParser);
         $extractor = new DependencyInjectionDocumentExtractor(
             $yamlExtractor,
-            new XmlDependencyInjectionExtractor($converter),
+            new XmlDependencyInjectionExtractor($converter, new ParameterExpressionScanner()),
             $autowireExtractor,
             $classExtractor,
         );